requirements.txt
pdf2image>=1.16.0
Pillow>=9.0.0
pytesseract>=0.3.8
PyMuPDF>=1.19.0
tqdm>=4.62.0
